Removed duplicated deprecated classes. Submenu moved to menu pages folder
